third_party/blink/web_tests/plugins/navigator-pluginarray.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function () {
  assert_equals(navigator.plugins.length, 0, "Plugins array must be empty");
  assert_equals(navigator.mimeTypes.length, 0, "mimeTypes array must be empty");
  navigator.plugins.refresh();
  assert_equals(navigator.plugins.length, 0, "Plugins array must be empty, even after refresh()");
  assert_equals(navigator.mimeTypes.length, 0, "mimeTypes array must be empty, even after refresh()");
}, "Tests that navigator.plugins and mimeTypes both return empty arrays.");
</script>
